requirements.txt
pymongo
kafka-python 
stop-words
